refactor: evaluators out of skills
