+++ src/catch/difficulty/evaluators/mod.rs
@@ -1,0 +1,3 @@
+pub use self::movement::MovementEvaluator;
+
+mod movement;
