+++ src/osu/difficulty/evaluators/mod.rs
@@ -1,0 +1,9 @@
+pub use self::{
+    aim::AimEvaluator, flashlight::FlashlightEvaluator, rhythm::RhythmEvaluator,
+    speed::SpeedEvaluator,
+};
+
+mod aim;
+mod flashlight;
+mod rhythm;
+mod speed;
