+++ src/osu/difficulty/skills/aim.rs
@@ -3,12 +3,8 @@ use crate::{
         object::{HasStartTime, IDifficultyObject},
         skills::{StrainSkill, strain_decay},
     },
-    osu::difficulty::object::OsuDifficultyObject,
-    util::{
-        difficulty::{milliseconds_to_bpm, reverse_lerp, smootherstep, smoothstep},
-        float_ext::FloatExt,
-        strains_vec::StrainsVec,
-    },
+    osu::difficulty::{evaluators::AimEvaluator, object::OsuDifficultyObject},
+    util::{float_ext::FloatExt, strains_vec::StrainsVec},
 };
 
 use super::strain::OsuStrainSkill;
@@ -90,224 +86,3 @@ impl Aim {
 }
 
 impl OsuStrainSkill for Aim {}
-
-struct AimEvaluator;
-
-impl AimEvaluator {
-    const WIDE_ANGLE_MULTIPLIER: f64 = 1.5;
-    const ACUTE_ANGLE_MULTIPLIER: f64 = 2.55;
-    const SLIDER_MULTIPLIER: f64 = 1.35;
-    const VELOCITY_CHANGE_MULTIPLIER: f64 = 0.75;
-    const WIGGLE_MULTIPLIER: f64 = 1.02;
-
-    #[expect(clippy::too_many_lines, reason = "staying in-sync with lazer")]
-    fn evaluate_diff_of<'a>(
-        curr: &'a OsuDifficultyObject<'a>,
-        diff_objects: &'a [OsuDifficultyObject<'a>],
-        with_slider_travel_dist: bool,
-    ) -> f64 {
-        let osu_curr_obj = curr;
-
-        let Some((osu_last_last_obj, osu_last_obj)) = curr
-            .previous(1, diff_objects)
-            .zip(curr.previous(0, diff_objects))
-            .filter(|(_, last)| !(curr.base.is_spinner() || last.base.is_spinner()))
-        else {
-            return 0.0;
-        };
-
-        #[expect(clippy::items_after_statements, reason = "staying in-sync with lazer")]
-        const RADIUS: i32 = OsuDifficultyObject::NORMALIZED_RADIUS;
-        #[expect(clippy::items_after_statements, reason = "staying in-sync with lazer")]
-        const DIAMETER: i32 = OsuDifficultyObject::NORMALIZED_DIAMETER;
-
-        // * Calculate the velocity to the current hitobject, which starts
-        // * with a base distance / time assuming the last object is a hitcircle.
-        let mut curr_vel = osu_curr_obj.lazy_jump_dist / osu_curr_obj.adjusted_delta_time;
-
-        // * But if the last object is a slider, then we extend the travel
-        // * velocity through the slider into the current object.
-        if osu_last_obj.base.is_slider() && with_slider_travel_dist {
-            // * calculate the slider velocity from slider head to slider end.
-            let travel_vel = osu_last_obj.travel_dist / osu_last_obj.travel_time;
-            // * calculate the movement velocity from slider end to current object
-            let movement_vel = osu_curr_obj.min_jump_dist / osu_curr_obj.min_jump_time;
-
-            // * take the larger total combined velocity.
-            curr_vel = curr_vel.max(movement_vel + travel_vel);
-        }
-
-        // * As above, do the same for the previous hitobject.
-        let mut prev_vel = osu_last_obj.lazy_jump_dist / osu_last_obj.adjusted_delta_time;
-
-        if osu_last_last_obj.base.is_slider() && with_slider_travel_dist {
-            let travel_vel = osu_last_last_obj.travel_dist / osu_last_last_obj.travel_time;
-            let movement_vel = osu_last_obj.min_jump_dist / osu_last_obj.min_jump_time;
-
-            prev_vel = prev_vel.max(movement_vel + travel_vel);
-        }
-
-        let mut wide_angle_bonus = 0.0;
-        let mut acute_angle_bonus = 0.0;
-        let mut slider_bonus = 0.0;
-        let mut vel_change_bonus = 0.0;
-        let mut wiggle_bonus = 0.0;
-
-        // * Start strain with regular velocity.
-        let mut aim_strain = curr_vel;
-
-        if let Some((curr_angle, last_angle)) = osu_curr_obj.angle.zip(osu_last_obj.angle) {
-            // * Rewarding angles, take the smaller velocity as base.
-            let angle_bonus = curr_vel.min(prev_vel);
-
-            // * If rhythms are the same.
-            if osu_curr_obj
-                .adjusted_delta_time
-                .max(osu_last_obj.adjusted_delta_time)
-                < 1.25
-                    * osu_curr_obj
-                        .adjusted_delta_time
-                        .min(osu_last_obj.adjusted_delta_time)
-            {
-                acute_angle_bonus = Self::calc_acute_angle_bonus(curr_angle);
-
-                // * Penalize angle repetition.
-                acute_angle_bonus *= 0.08
-                    + 0.92
-                        * (1.0
-                            - f64::min(
-                                acute_angle_bonus,
-                                f64::powf(Self::calc_acute_angle_bonus(last_angle), 3.0),
-                            ));
-
-                // * Apply acute angle bonus for BPM above 300 1/2 and distance more than one diameter
-                acute_angle_bonus *= angle_bonus
-                    * smootherstep(
-                        milliseconds_to_bpm(osu_curr_obj.adjusted_delta_time, Some(2)),
-                        300.0,
-                        400.0,
-                    )
-                    * smootherstep(
-                        osu_curr_obj.lazy_jump_dist,
-                        f64::from(DIAMETER),
-                        f64::from(DIAMETER * 2),
-                    );
-            }
-
-            wide_angle_bonus = Self::calc_wide_angle_bonus(curr_angle);
-
-            // * Penalize angle repetition.
-            wide_angle_bonus *= 1.0
-                - f64::min(
-                    wide_angle_bonus,
-                    f64::powf(Self::calc_wide_angle_bonus(last_angle), 3.0),
-                );
-
-            // * Apply full wide angle bonus for distance more than one diameter
-            wide_angle_bonus *=
-                angle_bonus * smootherstep(osu_curr_obj.lazy_jump_dist, 0.0, f64::from(DIAMETER));
-
-            // * Apply wiggle bonus for jumps that are [radius, 3*diameter] in distance, with < 110 angle
-            // * https://www.desmos.com/calculator/dp0v0nvowc
-            wiggle_bonus = angle_bonus
-                * smootherstep(
-                    osu_curr_obj.lazy_jump_dist,
-                    f64::from(RADIUS),
-                    f64::from(DIAMETER),
-                )
-                * f64::powf(
-                    reverse_lerp(
-                        osu_curr_obj.lazy_jump_dist,
-                        f64::from(DIAMETER * 3),
-                        f64::from(DIAMETER),
-                    ),
-                    1.8,
-                )
-                * smootherstep(curr_angle, f64::to_radians(110.0), f64::to_radians(60.0))
-                * smootherstep(
-                    osu_last_obj.lazy_jump_dist,
-                    f64::from(RADIUS),
-                    f64::from(DIAMETER),
-                )
-                * f64::powf(
-                    reverse_lerp(
-                        osu_last_obj.lazy_jump_dist,
-                        f64::from(DIAMETER * 3),
-                        f64::from(DIAMETER),
-                    ),
-                    1.8,
-                )
-                * smootherstep(last_angle, f64::to_radians(110.0), f64::to_radians(60.0));
-
-            if let Some(osu_last_2_obj) = curr.previous(2, diff_objects) {
-                let distance = f64::from(
-                    (osu_last_2_obj.base.stacked_pos() - osu_last_obj.base.stacked_pos()).length(),
-                );
-
-                if distance < 1.0 {
-                    wide_angle_bonus *= 1.0 - 0.35 * (1.0 - distance);
-                }
-            }
-        }
-
-        if prev_vel.max(curr_vel).not_eq(0.0) {
-            // * We want to use the average velocity over the whole object when awarding
-            // * differences, not the individual jump and slider path velocities.
-            prev_vel = (osu_last_obj.lazy_jump_dist + osu_last_last_obj.travel_dist)
-                / osu_last_obj.adjusted_delta_time;
-            curr_vel = (osu_curr_obj.lazy_jump_dist + osu_last_obj.travel_dist)
-                / osu_curr_obj.adjusted_delta_time;
-
-            // * Scale with ratio of difference compared to 0.5 * max dist.
-            let dist_ratio = smoothstep(
-                (prev_vel - curr_vel).abs() / prev_vel.max(curr_vel),
-                0.0,
-                1.0,
-            );
-
-            // * Reward for % distance up to 125 / strainTime for overlaps where velocity is still changing.
-            let overlap_vel_buff = (f64::from(DIAMETER) * 1.25
-                / osu_curr_obj
-                    .adjusted_delta_time
-                    .min(osu_last_obj.adjusted_delta_time))
-            .min((prev_vel - curr_vel).abs());
-
-            vel_change_bonus = overlap_vel_buff * dist_ratio;
-
-            // * Penalize for rhythm changes.
-            let bonus_base = (osu_curr_obj.adjusted_delta_time)
-                .min(osu_last_obj.adjusted_delta_time)
-                / (osu_curr_obj.adjusted_delta_time).max(osu_last_obj.adjusted_delta_time);
-            vel_change_bonus *= bonus_base.powf(2.0);
-        }
-
-        if osu_last_obj.base.is_slider() {
-            // * Reward sliders based on velocity.
-            slider_bonus = osu_last_obj.travel_dist / osu_last_obj.travel_time;
-        }
-
-        aim_strain += wiggle_bonus * Self::WIGGLE_MULTIPLIER;
-        aim_strain += vel_change_bonus * Self::VELOCITY_CHANGE_MULTIPLIER;
-
-        // * Add in acute angle bonus or wide angle bonus, whichever is larger.
-        aim_strain += (acute_angle_bonus * Self::ACUTE_ANGLE_MULTIPLIER)
-            .max(wide_angle_bonus * Self::WIDE_ANGLE_MULTIPLIER);
-
-        aim_strain *= osu_curr_obj.small_circle_bonus;
-
-        // * Add in additional slider velocity bonus.
-        if with_slider_travel_dist {
-            aim_strain += slider_bonus * Self::SLIDER_MULTIPLIER;
-        }
-
-        aim_strain
-    }
-
-    const fn calc_wide_angle_bonus(angle: f64) -> f64 {
-        smoothstep(angle, f64::to_radians(40.0), f64::to_radians(140.0))
-    }
-
-    const fn calc_acute_angle_bonus(angle: f64) -> f64 {
-        smoothstep(angle, f64::to_radians(140.0), f64::to_radians(40.0))
-    }
-}
