+++ src/taiko/difficulty/evaluators/mod.rs
@@ -1,0 +1,9 @@
+pub use self::{
+    color::ColorEvaluator, reading::ReadingEvaluator, rhythm::RhythmEvaluator,
+    stamina::StaminaEvaluator,
+};
+
+mod color;
+mod reading;
+mod rhythm;
+mod stamina;
